feat add product component
